Fix LiteParse review findings: drop numpy import, log dropped images, dedupe annotation builders
- opencontractserver/pipeline/parsers/liteparse_parser.py and
  llamaparse_parser.py: stop importing numpy solely for a
  dict[int, np.ndarray] type annotation. numpy is a transitive
  dependency (via shapely et al.), not a declared hard requirement, so
  an environment without it would fail to import the module entirely,
  blocking pipeline registry discovery of these parsers at startup.
  Annotation changed to dict[int, Any].
- liteparse_parser.py::_append_image_tokens: log a warning when
  embedded images are dropped because their page index falls outside
  the parsed PAWLs pages (e.g. target_pages restricting text parsing
  to a narrower range than the full PDF). Previously this was silent
  data loss. Added a regression test asserting the warning fires.
- Lift _create_annotation() and _build_image_token() out of
  LlamaParseParser and LiteParseParser into BaseParser
  (opencontractserver/pipeline/base/parser.py) so the two near-verbatim
  copies aren't maintained independently; liteparse_parser.py's
  image-token construction now calls the shared _build_image_token
  instead of re-deriving the same dict inline.
- Remove the redundant DEFAULT_WIDTH/DEFAULT_HEIGHT module-level
  aliases in liteparse_parser.py in favor of the canonical
  DEFAULT_PDF_PAGE_WIDTH/DEFAULT_PDF_PAGE_HEIGHT constants used
  everywhere else in the file.
- Move the _attr() dict-or-dataclass helper out of liteparse_parser.py
  into opencontractserver/utils/object_access.py so it lives in the
  shared utils package and other parsers can reuse it.
